backend/requirements.txt
Flask
rembg
onnxruntime
numpy==1.26.0
transformers==4.47.1
diffusers
torch
pillow
accelerate

